Rename useInitCaptcha file
